+++ src/index.ts
@@ -39,7 +39,7 @@ import {
 import { Vault } from "./vault.js";
 import { VaultWatcher } from "./watcher.js";
 
-const VERSION = "1.11.0";
+const VERSION = "1.11.1";
 
 interface ServeOptions {
   vault: string;